Refactor SignUpPage component filename and import statement
